requirements.txt
requests
requests_toolbelt
beautifulsoup4
GoogleAppEngineCloudStorageClient==1.9.22.1
